RW-5796 create test class for criteria search page
